Allow building with Clang using MSVC ABI
The official LLVM build of Clang is using the MSVC ABI, and
if you are using clang.exe and not clang-cl.exe you get the
GNU command line interface.

CMake would not set the MSVC variable, but the CMAKE_CXX_SIMULATE_ID
as "MSVC". Since clang.exe is targeting the MSVC ABI it needs
the build support the MSVC has.